Move date_range_or_funds_path task to digital_bookplate_instances
