+++ libsys_airflow/plugins/folio/invoices.py
@@ -167,28 +167,3 @@ def invoice_lines_paid_on_fund(**kwargs) -> list:
             logger.info(f"Fund does not have fund_uuid: {row}")
 
     return all_invoice_lines
-
-
-@task.branch()
-def date_range_or_funds_path(**kwargs):
-    # params["funds"]:
-    """
-    [
-        {
-            'druid': 'ef919yq2614',
-            'failure': None,
-            'fund_name': 'KELP',
-            'title': 'The Kelp Foundation Fund',
-            'image_filename': 'ef919yq2614_00_0001.jp2',
-            'db_id': 4,
-            'fund_uuid': 'f916c6e4-1bc7-4892-a5a8-73b8ede6e3a4'
-        },
-    ]
-    """
-    params = kwargs.get("params", {})
-    funds = params.get("funds", [])
-    if len(funds) > 0:
-        logger.info(f"New fund {funds}")
-        return "invoice_lines_paid_on_fund"
-    else:
-        return "invoices_paid_within_date_range"
